+++ sdk/go/random/randomPet.go
@@ -7,11 +7,11 @@ import (
 	"github.com/pulumi/pulumi/sdk/go/pulumi"
 )
 
-// The resource `random_pet` generates random pet names that are intended to be
+// The resource `.RandomPet` generates random pet names that are intended to be
 // used as unique identifiers for other resources.
 // 
 // This resource can be used in conjunction with resources that have
-// the `create_before_destroy` lifecycle flag set, to avoid conflicts with
+// the `createBeforeDestroy` lifecycle flag set, to avoid conflicts with
 // unique names during the brief period where both the old and new resources
 // exist concurrently.
 //