+++ sdk/python/pulumi_random/random_uuid.py
@@ -6,6 +6,7 @@ import json
 import warnings
 import pulumi
 import pulumi.runtime
+from typing import Union
 from . import utilities, tables
 
 class RandomUuid(pulumi.CustomResource):
@@ -19,9 +20,9 @@ class RandomUuid(pulumi.CustomResource):
     """
     The generated uuid presented in string format.
     """
-    def __init__(__self__, resource_name, opts=None, keepers=None, __name__=None, __opts__=None):
+    def __init__(__self__, resource_name, opts=None, keepers=None, __props__=None, __name__=None, __opts__=None):
         """
-        The resource `random_uuid` generates random uuid string that is intended to be
+        The resource `.RandomUuid` generates random uuid string that is intended to be
         used as unique identifiers for other resources.
         
         This resource uses the `hashicorp/go-uuid` to generate a UUID-formatted string
@@ -41,26 +42,47 @@ class RandomUuid(pulumi.CustomResource):
         if __opts__ is not None:
             warnings.warn("explicit use of __opts__ is deprecated, use 'opts' instead", DeprecationWarning)
             opts = __opts__
-        if not resource_name:
-            raise TypeError('Missing resource name argument (for URN creation)')
-        if not isinstance(resource_name, str):
-            raise TypeError('Expected resource name to be a string')
-        if opts and not isinstance(opts, pulumi.ResourceOptions):
+        if opts is None:
+            opts = pulumi.ResourceOptions()
+        if not isinstance(opts, pulumi.ResourceOptions):
             raise TypeError('Expected resource options to be a ResourceOptions instance')
+        if opts.version is None:
+            opts.version = utilities.get_version()
+        if opts.id is None:
+            if __props__ is not None:
+                raise TypeError('__props__ is only valid when passed in combination with a valid opts.id to get an existing resource')
+            __props__ = dict()
 
-        __props__ = dict()
-
-        __props__['keepers'] = keepers
-
-        __props__['result'] = None
-
+            __props__['keepers'] = keepers
+            __props__['result'] = None
         super(RandomUuid, __self__).__init__(
             'random:index/randomUuid:RandomUuid',
             resource_name,
             __props__,
             opts)
 
+    @staticmethod
+    def get(resource_name, id, opts=None, keepers=None, result=None):
+        """
+        Get an existing RandomUuid resource's state with the given name, id, and optional extra
+        properties used to qualify the lookup.
+        
+        :param str resource_name: The unique name of the resulting resource.
+        :param str id: The unique provider ID of the resource to lookup.
+        :param pulumi.ResourceOptions opts: Options for the resource.
+        :param pulumi.Input[dict] keepers: Arbitrary map of values that, when changed, will
+               trigger a new uuid to be generated. See
+               the main provider documentation for more information.
+        :param pulumi.Input[str] result: The generated uuid presented in string format.
 
+        > This content is derived from https://github.com/terraform-providers/terraform-provider-random/blob/master/website/docs/r/uuid.html.markdown.
+        """
+        opts = pulumi.ResourceOptions.merge(opts, pulumi.ResourceOptions(id=id))
+
+        __props__ = dict()
+        __props__["keepers"] = keepers
+        __props__["result"] = result
+        return RandomUuid(resource_name, opts=opts, __props__=__props__)
     def translate_output_property(self, prop):
         return tables._CAMEL_TO_SNAKE_CASE_TABLE.get(prop) or prop